test: extract shared HTTP helpers into OpenAiServerTestSupport base
Both OpenAiCompatServerHttpTest and OpenAiCompatServerIntegrationTest duplicated
the post/get/read/readAll helpers and the Response holder. Move them into a new
abstract OpenAiServerTestSupport (intentionally not named *Test, so the harness
never runs it on its own); both classes now extend it and call the inherited
post(port, path, body, auth) / get(port, path, auth).

Behaviour is unchanged: the fake-backend HTTP tests pass and the model-gated
integration test still self-skips when the model is absent.
