+++ src/test/java/net/ladenthin/llama/server/OpenAiCompatServerIntegrationTest.java
@@ -4,7 +4,6 @@
 
 package net.ladenthin.llama.server;
 
-import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.greaterThan;
@@ -12,13 +11,8 @@ import static org.hamcrest.Matchers.is;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
-import java.net.HttpURLConnection;
-import java.net.URL;
 import net.ladenthin.llama.LlamaModel;
 import net.ladenthin.llama.TestConstants;
 import net.ladenthin.llama.parameters.ModelParameters;
@@ -37,9 +31,10 @@ import org.junit.jupiter.api.Test;
  * <p>Assertions are deliberately structural (valid OpenAI shapes, stream terminates) rather than
  * content-specific — a 0.6B model's exact wording and whether it elects to call a tool are not
  * deterministic. The deterministic chunk/tool-call plumbing is covered by
- * {@link OpenAiCompatServerHttpTest} with a fake backend.
+ * {@link OpenAiCompatServerHttpTest} with a fake backend. HTTP request plumbing is inherited from
+ * {@link OpenAiServerTestSupport}.
  */
-public class OpenAiCompatServerIntegrationTest {
+public class OpenAiCompatServerIntegrationTest extends OpenAiServerTestSupport {
 
     private static final ObjectMapper MAPPER = new ObjectMapper();
     private static final String MODEL_ID = "qwen3-local";
@@ -81,7 +76,7 @@ public class OpenAiCompatServerIntegrationTest {
     public void nonStreamingChatReturnsValidCompletion() throws IOException {
         String body = "{\"model\":\"" + MODEL_ID + "\",\"max_tokens\":16,"
                 + "\"messages\":[{\"role\":\"user\",\"content\":\"Say hello in one word.\"}]}";
-        Response response = post("/v1/chat/completions", body);
+        Response response = post(port, "/v1/chat/completions", body, "");
         assertThat(response.code, is(200));
         JsonNode json = MAPPER.readTree(response.body);
         assertThat(json.path("object").asText(), is("chat.completion"));
@@ -93,7 +88,7 @@ public class OpenAiCompatServerIntegrationTest {
     public void streamingChatEmitsChunksAndDone() throws IOException {
         String body = "{\"model\":\"" + MODEL_ID + "\",\"stream\":true,\"max_tokens\":16,"
                 + "\"messages\":[{\"role\":\"user\",\"content\":\"Say hello in one word.\"}]}";
-        Response response = post("/v1/chat/completions", body);
+        Response response = post(port, "/v1/chat/completions", body, "");
         assertThat(response.code, is(200));
         assertThat(response.body, containsString("chat.completion.chunk"));
         assertThat(response.body, containsString("data: [DONE]"));
@@ -109,7 +104,7 @@ public class OpenAiCompatServerIntegrationTest {
                 + "\"tools\":[{\"type\":\"function\",\"function\":{\"name\":\"get_weather\","
                 + "\"description\":\"Get the weather for a city\",\"parameters\":{\"type\":\"object\","
                 + "\"properties\":{\"city\":{\"type\":\"string\"}},\"required\":[\"city\"]}}}]}";
-        Response response = post("/v1/chat/completions", body);
+        Response response = post(port, "/v1/chat/completions", body, "");
         assertThat(response.code, is(200));
         JsonNode message = MAPPER.readTree(response.body).path("choices").get(0).path("message");
         assertThat(message.isObject(), is(true));
@@ -117,55 +112,8 @@ public class OpenAiCompatServerIntegrationTest {
 
     @Test
     public void modelsEndpointAdvertisesTheServedModel() throws IOException {
-        Response response = get("/v1/models");
+        Response response = get(port, "/v1/models", "");
         assertThat(response.code, is(200));
         assertThat(response.body, containsString(MODEL_ID));
-    }
-
-    // ----- HTTP helpers -----
-
-    private static Response post(String path, String body) throws IOException {
-        HttpURLConnection conn = (HttpURLConnection) new URL("http://127.0.0.1:" + port + path).openConnection();
-        conn.setRequestMethod("POST");
-        conn.setDoOutput(true);
-        conn.setRequestProperty("Content-Type", "application/json");
-        try (OutputStream os = conn.getOutputStream()) {
-            os.write(body.getBytes(UTF_8));
-        }
-        return read(conn);
-    }
-
-    private static Response get(String path) throws IOException {
-        HttpURLConnection conn = (HttpURLConnection) new URL("http://127.0.0.1:" + port + path).openConnection();
-        conn.setRequestMethod("GET");
-        return read(conn);
-    }
-
-    private static Response read(HttpURLConnection conn) throws IOException {
-        int code = conn.getResponseCode();
-        InputStream is = code < 400 ? conn.getInputStream() : conn.getErrorStream();
-        String body = is == null ? "" : readAll(is);
-        return new Response(code, body);
-    }
-
-    private static String readAll(InputStream is) throws IOException {
-        ByteArrayOutputStream buffer = new ByteArrayOutputStream();
-        byte[] chunk = new byte[1024];
-        int read;
-        while ((read = is.read(chunk)) != -1) {
-            buffer.write(chunk, 0, read);
-        }
-        return new String(buffer.toByteArray(), UTF_8);
-    }
-
-    /** Minimal HTTP response holder. */
-    private static final class Response {
-        private final int code;
-        private final String body;
-
-        Response(int code, String body) {
-            this.code = code;
-            this.body = body;
-        }
     }
 }
